Playtest additions and fixes (#10936)
* HuD & AdminTools

* More Fix

* BuckledRestrictions

* Closing Some Issues

* Tests

+++ UnityProject/Assets/Scripts/Systems/InGameEvents/SicknessEventParameters.cs
@@ -19,6 +19,6 @@
 		/// <summary>
 		/// The starting % of the disease in applied victims
 		/// </summary>
-		public int Strength;
+		public float Strength;
 	}
 }
